Adding id into to all structs after creation in db

+++ api/service/handler.go
@@ -9,7 +9,7 @@ import (
 func CreateServiceHandler(w http.ResponseWriter, r *http.Request) {
 	appId, _ := strconv.Atoi(r.FormValue("ServiceTypeId"))
 	service := Service{
-		ServiceTypeId:   appId,
+		ServiceTypeId:   int64(appId),
 		Name:            r.FormValue("name"),
 	}
